chore: rename statistics because of typo
