+++ src/main/java/org/pedrocarlos/Main.java
@@ -3,8 +3,8 @@ package org.pedrocarlos;
 import java.util.Arrays;
 import java.util.List;
 import org.pedrocarlos.core.FileTransformerExecutor;
-import org.pedrocarlos.statitics.Statistics;
-import org.pedrocarlos.statitics.ReadAndWriteStreamWithStatistics;
+import org.pedrocarlos.statistics.Statistics;
+import org.pedrocarlos.statistics.ReadAndWriteStreamWithStatistics;
 import org.pedrocarlos.app.FileTransformerCommandLine;
 import org.pedrocarlos.core.operations.CapitalizeTransformer;
 import org.pedrocarlos.core.operations.NegativeTransformer;
